requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
Pillow
python-dotenv
